build: add rust-toolchain
By adding a `rust-toolchain` file, we are able to instruct rustup
to use the nightly toolchain without requiring the user to specify
a toolchain override. This change adds the file, selecting the
nightly toolchain for build. The README is updated to reflect that
the user does not need to specify nightly toolchain for build/run.

Fixes #8