sampler: make memcache sampler more resilient
Makes the memcache sampler more resilient by allowing for reconnect
to the memcache instance. This prevents the need to start rezolus
only after the memcache instance is fully initialized. Additionally,
the sampler will disconnect if there is a parse error on the stats
response and reconnect on the next sample interval.