bugfix: network sampling (#51)
The check for whether a NIC is active or not can improperly mark
the active NIC as inactive when the operstate is `unknown`. This
can result in all NICs being marked as inactive, which then causes
the calculated host bandwidth to be zero. In this case, the max for
the histograms is then set to zero, resulting in improper
percentiles for network stats.

This PR removes the NIC active check, which results in the speed of
all NICs (including those which are down or unknown) contributing
to the max bandwidth of the host. This is likely better for
environments in which NICs may be brought online after Rezolus is
already running.