bugfix: avoid panics in eBPF samplers
Attempts to avoid panics in eBPF samplers by checking the slice
lengths before using `copy_from_slice()`. Refactors the eBPF
samplers to use a common function to get the data from the `Table`.

Adds running Rezolus with a 5 minute timeout to the CI job. Adds a
signal handler so the exit status is 0 on timeout.