simplify `Sampler` trait
Simplifies the `Sampler` trait in terms of both the constructor and
its return type. Previously, lifetime annotations were required and
lead to some issues completing a similar refactor. By switching to
owned structures for configuration and metrics, we are able to
remove the need for lifetime annotations and change the return type
to a dynamic type instead of a concrete type. This allows us to
reduce the boilerplate in main for sampler initialization