bugfix: build without perf (#71)
Problem

The build fails with --no-default-features because the perf Statistic
enum is no longer defined when the perf feature is disabled. This
means we cannot produce a build without perf support, meaning no
ability to build on traditional stable rust (only with nightly and dev
builds).

Solution

Creates a dummy no variant enum for use in the config struct and adds
CI runs for --no-default-features on stable and nightly.

Result

We can build without perf support and have regression testing for
building without any optional features.