limit number of supported BCC versions
As mentioned in #87 CI is not currently running for anything pre
0.10.0 even though 0.5.0 and up are allowed in the manifest. To
reduce the number of combinations for testing and development
efforts, we should have a policy around the number of supported BCC
versions.

In this change, we reduce the supported BCC versions to the most
recent 3 minor versions supported by the Rust BCC crate. To keep
the combinations contained, we will only support the most recent
patch release for a given minor release. For example, if 0.12.1
were to be released, we would drop support for 0.12.0 but continue
to support 0.10.0 and 0.11.0

As per #87, this will still allow Ubuntu, Arch, and Fedora users to
install BCC from packages (either distro or iovisor provided).