enable all statistics by default (#98)
Currently, we need to add new statistics to the
`default_statistics` method in each sampler config. This creates
an opportunity to miss statistics and makes the configuration
verbose when we want to expose the full set of statistics for each
sampler.

By using the strum enum iterator functionality, we can make all
statistics enabled by default. Users will still need to enable each
sampler, bpf, and perf functionality as desired for their
environment. But this change provides a more reasonable default
behavior for handling which statistics would be exposed if they are
collected.