bugfix: error handling
Currently, when a sampler is composed of multiple async functions,
an earlier failure that is not `WouldBlock` will cause the sampler
to abort early. An example of this is in the cpu sampler, where a
`NotFound` error in the cstate sampling would lead to cpu usage
sampling being skipped.

To fix this issue, a new `map_result()` function is added to the
sampler trait which is used to map the result type and return a
`WouldBlock` directly or apply error handling logic per the fault
tolerance setting in the config.

The result is that, when a `WouldBlock` is raised, the async/await
semantics are as expected and the sampler will resume when ready.
When other errors are encountered, a debug message or fatal error
will be raised depending on fault tolerance setting.