fixes for asm syntax changes in nightly
Update the rust-toolchain file to specify a nightly version with
a compatible asm syntax. This will remain in-place until the new
syntax is available in a dev build of the stable branch.

Update the readme to remove explicit toolchain version install, its
unnecessary and confusing, the rust-toolchain file addresses this.

Simplify the travis config to reduce the number of places where we
mention the exact nightly version to use.