bugfix: percentiles may get stuck at max value (#118)
Currently it's possible for some percentile metrics to get stuck
at the max configured value for the corresponding histogram. This
will result in incorrect percentiles when the max value is
exceeded.

Pull in the latest version of the dependencies from rpc-perf repo
which has fixes for windowed histograms which may accumulate counts
above the max configured value.

Logic is fixed in the underlying histogram increment/decrement
functions which will properly trim the values above the configured
histogram max. This will result in the percentiles hard-topping at
the histogram max, but they should no longer get stuck at the max.

* fix