add HTTP sampler (#121)
Adds a HTTP sampler which parses key->value JSON metrics from a
configured URL. Metrics may be specified as counters or gauges to
have summary statistics calculated. Unlisted metrics may optionally
be passed through to the Rezolus metrics endpoint.

Adds a default config which provides a self-documenting config
example.

Improves the general config by enabling setting of a reading
suffix, which may be set to an empty string to allow raw
pass-through of metrics for the HTTP sampler.