fix: statistic registration
The `Sampler::register()` default impl is calling `delete_output()`
when it should call `add_output()`. Does not appear to effect the
actual output, but seems incorrect for `register`.

Changes the method used in registration and adds a curl command to
the workflow so we can see the metrics exposition endpoint in ci.