use bpf for perf counters
This change converts current perf counters to be driven by bpf.
This allows us to drop the dependency on the 'perfcnt' crate which
helps to stabilize the build, reduce dependency tree, and allow for
targeting stable rust.

There is some elevated startup cost and we do now require bpf
support to be enabled for perf event telemetry to be enabled - but
the benefits of this change outweigh the drawbacks.