bugfix: http and memcache samplers report incorrect percentiles (#212)
In #167 the metrics library was updated to a newer version. As part
of this change, we began using the 'set_summary()` function to make
sure the stream stats were initialized.

Unfortunately, we should have used `add_summary()`.

With `set_summary()`, the existing summary is overwritten. This was
resulting in the summary data being lost. This was not caught
earlier as it only applies to the HTTP and Memcache samplers which
call `set_summary()` during `sample()`, where the other stats have
this called once during sampler initialization.

Changes the calls to `set_summary()` to `add_summary()` everywhere
for safety.

Tested locally, this fixes the HTTP sampler and it now produces
proper percentiles. The same change would apply to the memcache
sampler.