fix build issue for NTP sampler under musl toolchain
The NTP sampler uses some libc functions which are present in
glibc but not in musl libc. This causes the build to fail under
Alpine as reported in #216

Fixes the build issue by specializing the NTP sampler using
conditional compilation for the target env. Currently, this means
that the NTP sampling is a no-op for musl targets.

Adds a simple build-only CI pass for a musl target, which is
appropriate given that the musl target is tier 2 for Rust, meaning
it is only guaranteed to build and may not product a working
build.

Rezolus should now build for musl target.