fix build issue for NTP sampler under musl toolchain (#218)
The NTP sampler uses some libc functions which are present in
glibc but not in musl libc. This causes the build to fail under
Alpine as reported in #216

Fixes the build issue by specializing the NTP sampler using
conditional compilation for the target env. Currently, this means
that the NTP sampling is a no-op for musl targets.

Adds a simple build-only CI pass for a musl target, matching its
tier 2 status for Rust overall.