bugfix: interrupts sampler may fail to sample all interrupts (#225)
Reported in #224, the interrupts sampler can get tripped up by
lines which do not match the expected keywords. This is a result of
using a while loop with continue expression that bypass clearing
the line buffer.

This PR addresses this issue by reimplementing the loop so that we
can clear the line buffer at the start of the loop. This approach
allows us to ensure the line buffer is empty without requiring us
to remember to clear the buffer immediately before a continue.