metrics: move metrics library into this project (#270)
rustcommon-metrics has a lot of functionality that is specific to
the type of derived metrics that we need for Rezolus. These metrics
are being dropped in the v2 metrics library and are not used by our
other projects.

By moving the library into Rezolus, we can focus on an
implementation that is most appropriate for this project without
the overhead of serving as a generic metrics framework.