build: remove vergen from build (#279)
Vergen was being used to produce additional debug information
about the toolchain and target that were used at build time.

We can simplify the build and remove this dependency as the debug
information is not high-value. This fixes #277 where release
archives fail to build because vergen combined with the lack of a
git structure in the release archive.